+++ config/opal_config_files.m4
@@ -4,6 +4,7 @@
 # Copyright (c) 2020      The University of Tennessee and The University
 #                         of Tennessee Research Foundation.  All rights
 #                         reserved.
+# Copyright (c) 2022      Amazon.com, Inc. or its affiliates.  All Rights reserved.
 # $COPYRIGHT$
 #
 # Additional copyrights may follow
@@ -16,6 +17,7 @@
 AC_DEFUN([OPAL_CONFIG_FILES],[
     AC_CONFIG_FILES([
         opal/Makefile
+        opal/cuda/Makefile
         opal/etc/Makefile
         opal/include/Makefile
         opal/datatype/Makefile
