+++ opal/cuda/Makefile.am
@@ -1,0 +1,44 @@
+#
+# Copyright (c) 2004-2005 The Trustees of Indiana University and Indiana
+#                         University Research and Technology
+#                         Corporation.  All rights reserved.
+# Copyright (c) 2004-2013 The University of Tennessee and The University
+#                         of Tennessee Research Foundation.  All rights
+#                         reserved.
+# Copyright (c) 2004-2009 High Performance Computing Center Stuttgart,
+#                         University of Stuttgart.  All rights reserved.
+# Copyright (c) 2004-2005 The Regents of the University of California.
+#                         All rights reserved.
+# Copyright (c) 2011-2013 NVIDIA Corporation.  All rights reserved.
+# Copyright (c) 2014-2015 Cisco Systems, Inc.  All rights reserved.
+# Copyright (c) 2022      Amazon.com, Inc. or its affiliates.  All Rights reserved.
+# $COPYRIGHT$
+#
+# Additional copyrights may follow
+#
+# $HEADER$
+#
+
+AM_CPPFLAGS = $(common_cuda_CPPFLAGS)
+
+# Header files
+headers = \
+        common_cuda.h
+
+# Source files
+sources = \
+        common_cuda.c
+
+dist_opaldata_DATA = help-mpi-common-cuda.txt
+
+noinst_LTLIBRARIES = libopalcuda.la
+
+libopalcuda_la_SOURCES = $(headers) $(sources)
+libopalcuda_la_LDFLAGS =
+libopalcuda_la_LIBADD =
+
+# Conditionally install the header files
+if WANT_INSTALL_HEADERS
+opaldir = $(opalincludedir)/$(subdir)
+opal_HEADERS = $(headers)
+endif
