+++ opal/cuda/common_cuda.c
